🐛 fix: missing kwargs in from_agent_and_tools in dataframe agent (#2285)
Hello! 
I've noticed a bug in `create_pandas_dataframe_agent`. When calling it
with argument `return_intermediate_steps=True`, it doesn't return the
intermediate step. I think the issue is that `kwargs` was not passed
where it needed to be passed. It should be passed into
`AgentExecutor.from_agent_and_tools`

Please correct me if my solution isn't appropriate and I will fix with
the appropriate approach.

Co-authored-by: alhajji <m.alhajji@drahim.sa>